refactor: 2주차 피드백 반영 및 axios 삭제
