+++ src/api/services/bookmarkService.js
@@ -1,8 +1,8 @@
-import axios from '../axios';
+import axiosConfig from '../axiosConfig';
 
 export const bookmarkService = {
     // 캠핑장 찜하기/취소하기 토글
     toggleBookmark: async (campId) => {
-        return axios.patch(`/api/camps/${campId}/bookmarks`);
+        return axiosConfig.patch(`/api/camps/${campId}/bookmarks`);
     }
 };
